requirements.txt
streamlit
pandas
google-generativeai
beautifulsoup4
requests
